+++ doc/conf.py
@@ -211,10 +211,14 @@ latex_elements = {
     # The font size ('10pt', '11pt' or '12pt').
     # 'pointsize': '10pt',
     # Additional stuff for the LaTeX preamble.
-    # 'preamble': '',
+    "preamble": r"""
+        \usepackage{amsthm}
+    """,
     # Latex figure (float) alignment
     # 'figure_align': 'htbp',
 }
+
+latex_proof_notheorem = ["proof"]
 
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title,