+++ sphinxcontrib/proof/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Louis Paternault
+# Copyright 2018 Louis Paternault
 #
 # Sphinxcontrib-Proof is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,24 +16,32 @@
 """Provide tools to typeset theorems, proofs, etc. in Sphinx documentation."""
 
 import os
-import pkg_resources
-
 from docutils import nodes
 from docutils.parsers.rst import directives, Directive
 
 from sphinx.util import copy_static_entry
 from sphinx.util.nodes import set_source_info
 
-
-def package_file(*filename):
-    """Return the path to a filename present in package data."""
-    return pkg_resources.resource_filename(
-        "sphinxcontrib.proof", os.path.join("data", *filename)
-    )
-
-
 VERSION = "1.0.1"
 PREFIX = "proof:"
+
+################################################################################
+# Configuration
+PROOF_THEOREM_LABELS = {
+    "algorithm": "Algorithm",
+    "conjecture": "Conjecture",
+    "corollary": "Corollary",
+    "definition": "Definition",
+    "example": "Example",
+    "lemma": "Lemma",
+    "observation": "Observation",
+    "proof": "Proof",
+    "property": "Property",
+    "theorem": "Theorem",
+}
+
+################################################################################
+# Docutils
 
 
 class StatementNode(nodes.General, nodes.Element):
@@ -42,46 +50,10 @@ class StatementNode(nodes.General, nodes.Element):
     pass
 
 
-class ProofNode(nodes.Part, nodes.Element):
-    """Proof"""
-
-    pass
-
-
 class ContentNode(nodes.General, nodes.Element):
     """Content of a proof or a statement"""
 
     pass
-
-
-class ProofEnvironment(Directive):
-    """A proof environment"""
-
-    has_content = True
-    required_arguments = 0
-    optional_arguments = 1
-    final_argument_whitespace = True
-    option_spec = {"label": directives.unchanged_required}
-
-    def run(self):
-        """Render this environment"""
-        self.env = self.state.document.settings.env
-        targetid = "index-%s" % self.env.new_serialno("index")
-        targetnode = nodes.target("", "", ids=[targetid])
-
-        node = ProofNode("\n".join(self.content))
-        node["classes"] += ["proof-type-proof"]
-
-        if self.arguments:
-            node["title"] = self.arguments[0]
-
-        content = ContentNode()
-        self.state.nested_parse(self.content, self.content_offset, content)
-        content["classes"] += ["proof-content"]
-        node += content
-
-        set_source_info(self, node)
-        return [targetnode, node]
 
 
 class StatementEnvironment(Directive):
@@ -114,33 +86,11 @@ class StatementEnvironment(Directive):
         return [targetnode, node]
 
 
+################################################################################
 # HTML
-def html_visit_proof_node(self, node):
-    """Enter :class:`ProofNode` in HTML builder."""
-    labels = parse_theorem_labels(self.builder.config)
-    self.body.append(self.starttag(node, "div"))
-    self.body.append("""<div class="proof-title">""")
-    self.body.append(
-        """<span class="proof-title-name">{}</span>""".format(
-            labels.get("proof", "Proof")
-        )
-    )
-    if "title" in node:
-        self.body.append("""<span class="proof-title-content">""")
-        self.body.append(u"({})".format(node["title"]))
-        self.body.append("""</span>""")
-    self.body.append("""</div>""")
-
-
-def html_depart_proof_node(self, node):
-    """Leave :class:`ProofNode` in HTML builder."""
-    # pylint: disable=unused-argument
-    self.body.append("</div>")
-
-
 def html_visit_statement_node(self, node):
     """Enter :class:`StatementNode` in HTML builder."""
-    labels = parse_theorem_labels(self.builder.config)
+    labels = self.builder.config.proof_theorem_labels
     self.body.append(self.starttag(node, "div"))
     self.body.append("""<div class="proof-title">""")
     self.body.append(
@@ -172,20 +122,8 @@ def html_depart_content_node(self, node):
     self.body.append("</div>")
 
 
+################################################################################
 # LaTeX
-def latex_visit_proof_node(self, node):
-    """Enter :class:`ProofNode` in LaTeX builder."""
-    self.body.append(r"\begin{proof}")
-    if "title" in node:
-        self.body.append("[{}]".format(node["title"]))
-    self.body.append("\n")
-
-
-def latex_depart_proof_node(self, node):
-    """Leave :class:`ProofNode` in LaTeX builder."""
-    # pylint: disable=unused-argument
-    self.body.append(r"\end{proof}")
-    self.body.append("\n")
 
 
 def latex_visit_statement_node(self, node):
@@ -214,53 +152,52 @@ def latex_depart_content_node(self, node):
     pass
 
 
-# latex sty not necessary anymore
-# user is still required to load desired theorem package
+def _latex_preamble_iterator(config):
+    labels = config.proof_theorem_labels
+
+    yield config.latex_elements.get("preamble", "")
+
+    yield r"\makeatletter"
+
+    if config.latex_proof_counter:
+        yield r"\newtheorem{%s}{%s}[%s]" % (
+            config.latex_proof_main,
+            labels[config.latex_proof_main],
+            config.latex_proof_counter,
+        )
+    else:
+        yield r"\newtheorem{%s}{%s}" % (
+            config.latex_proof_main,
+            labels[config.latex_proof_main],
+        )
+
+    for environment, label in labels.items():
+        if not environment in config.latex_proof_notheorem + [config.latex_proof_main]:
+            yield r"\newtheorem{%s}[%s]{%s}" % (
+                environment,
+                config.latex_proof_main,
+                label,
+            )
+
+    yield r"\makeatother"
+
+
+def latex_preamble(config):
+    """Return the custom LaTeX preamble."""
+    return "\n".join(_latex_preamble_iterator(config))
+
+
+################################################################################
+# Setup
+
+
 def builder_inited(app):
     """Hook called when builder has been inited."""
+    config = app.builder.config
     if app.builder.name == "latex":
-        counter = parse_theorem_counter(app.builder.config)
-        app.builder.config.latex_elements["preamble"] += "\n" + r"\makeatletter"
-        labels = parse_theorem_labels(app.builder.config)
-        app.builder.config.latex_elements[
-            "preamble"
-        ] += "\n" r"\newtheorem{theorem}{%s}[%s]" % (labels["theorem"], counter)
-        for environment, label in labels.items():
-            if not environment in ["proof", "theorem"]:
-                app.builder.config.latex_elements[
-                    "preamble"
-                ] += "\n" r"\newtheorem{%s}[theorem]{%s}" % (environment, label)
-
-        app.builder.config.latex_elements["preamble"] += "\n" + r"\makeatother"
-
-
-COUNTER = "proof_theorem_counter"
-
-THEOREMS = "proof_theorem_labels"
-
-
-def parse_theorem_counter(config):
-    return config.proof_theorem_counter
-
-
-def parse_theorem_labels(config):
-    return config.proof_theorem_labels
-
-
-# directives are built according to default labels
-# user can only overwrite existing labels
-DEFAULT_THEOREM_LABELS = {
-    "lemma": "Lemma",
-    "theorem": "Theorem",
-    "corollary": "Corollary",
-    "property": "Property",
-    "example": "Example",
-    "definition": "Definition",
-    "proof": "Proof",
-    "conjecture": "Conjecture",
-    "algorithm": "Algorithm",
-    "observation": "Observation",
-}
+        if "preamble" not in config.latex_elements:
+            config.latex_elements["preamble"] = ""
+        config.latex_elements["preamble"] += latex_preamble(config)
 
 
 def setup(app):
@@ -268,14 +205,11 @@ def setup(app):
     app.add_stylesheet("proof.css")
     app.add_javascript("proof.js")
 
-    app.add_config_value(THEOREMS, DEFAULT_THEOREM_LABELS, "env")
-    app.add_config_value(COUNTER, "chapter", "env")
+    app.add_config_value("proof_theorem_labels", PROOF_THEOREM_LABELS, "env")
+    app.add_config_value("latex_proof_counter", "", "env")
+    app.add_config_value("latex_proof_main", "theorem", "env")
+    app.add_config_value("latex_proof_notheorem", [], "env")
 
-    app.add_node(
-        ProofNode,
-        html=(html_visit_proof_node, html_depart_proof_node),
-        latex=(latex_visit_proof_node, latex_depart_proof_node),
-    )
     app.add_node(
         StatementNode,
         html=(html_visit_statement_node, html_depart_statement_node),
@@ -287,10 +221,7 @@ def setup(app):
         latex=(latex_visit_content_node, latex_depart_content_node),
     )
 
-    for environment in DEFAULT_THEOREM_LABELS:
-        if environment.strip() == "proof":
-            app.add_directive(PREFIX + environment, ProofEnvironment)
-        else:
-            app.add_directive(PREFIX + environment, StatementEnvironment)
+    for environment in app.config.proof_theorem_labels:
+        app.add_directive(PREFIX + environment, StatementEnvironment)
 
     app.connect("builder-inited", builder_inited)